Mundança para long int

+++ modulos/utilidades.h
@@ -1,6 +1,6 @@
 #ifndef UTILIDADES_H
 #define UTILIDADES_H
 
-int max(int a, int b);
+unsigned long int max(unsigned long int a, unsigned long int b);
 
 #endif
